@@ -8,57 +8,36 @@
     "watch": "ng build --watch --configuration development",
     "test": "ng test"
   },
-  "prettier": {
-    "overrides": [
-      {
-        "files": "*.html",
-        "options": {
-          "parser": "angular"
-        }
-      }
-    ]
-  },
   "private": true,
   "dependencies": {
     "@angular/animations": "^18.2.0",
-    "@angular/common": "^18.2.0",
+    "@angular/common": "^21.0.3",
+    "@angular/compiler": "^21.0.3",
     "@angular/core": "^18.2.0",
-    "@angular/fire": "^18.0.0",
-    "@angular/forms": "^18.2.0",
+    "@angular/fire": "^20.0.1",
+    "@angular/forms": "^21.0.3",
     "@angular/google-maps": "^18.0.0",
-    "@angular/platform-browser": "^18.2.0",
-    "@angular/platform-browser-dynamic": "^18.2.0",
-    "@angular/router": "^18.2.0",
+    "@angular/platform-browser": "^21.0.3",
+    "@angular/platform-browser-dynamic": "^21.0.3",
+    "@angular/router": "^21.0.3",
     "@googlemaps/markerclusterer": "^2.6.2",
     "@ngneat/hot-toast": "^6.1.0",
-    "firebase": "^12.6.0",
+    "firebase": "^10.12.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "^0.14.10"
+    "zone.js": "~0.14.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.21",
-    "@angular/animations": "^18.2.0",
-    "@angular/common": "^18.2.0",
-    "@angular/compiler": "^18.2.14",
-    "@angular/compiler-cli": "^18.2.14",
-    "@angular/core": "^18.2.0",
-    "@angular/fire": "^18.0.0",
-    "@angular/forms": "^18.2.0",
-    "@angular/google-maps": "^18.0.0",
-    "@angular/platform-browser": "^18.2.0",
-    "@angular/platform-browser-dynamic": "^18.2.0",
-    "@angular/router": "^18.2.0",
+    "@angular-devkit/build-angular": "^21.0.2",
+    "@angular/cli": "^21.0.2",
+    "@angular/compiler-cli": "^21.0.3",
     "@types/jasmine": "~5.1.0",
-    "autoprefixer": "^10.4.22",
     "jasmine-core": "~5.8.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "postcss": "^8.5.6",
-    "tailwindcss": "^3.4.10",
-    "typescript": "~5.5.4"
+    "typescript": "^5.9.3"
   }
 }
