@@ -1,20 +1,24 @@
-import { enableProdMode } from '@angular/core';
+// file: src/main.ts
 import { bootstrapApplication } from '@angular/platform-browser';
+import { importProvidersFrom } from '@angular/core';
 import { provideRouter } from '@angular/router';
-import { provideAnimations } from '@angular/platform-browser/animations';
-import { AppComponent } from './app/app.component';
-import { routes } from './app/app.routes';
-import 'zone.js';
+import { provideFirebaseApp, initializeApp } from '@angular/fire/app';
+import { provideAuth, getAuth } from '@angular/fire/auth';
+import { provideFirestore, getFirestore } from '@angular/fire/firestore';
+import { provideStorage, getStorage } from '@angular/fire/storage';
+import { HotToastModule } from '@ngneat/hot-toast';
 
-if (typeof window !== 'undefined' && (window as any).Cypress) {
-  // Cypress support
-} else if (location.hostname !== 'localhost' && location.hostname !== '127.0.0.1') {
-  enableProdMode();
-}
+import { routes } from './app/app.routes';
+import { AppComponent } from './app/app.component';
+import { environment } from './environments/environment';
 
 bootstrapApplication(AppComponent, {
   providers: [
     provideRouter(routes),
-    provideAnimations()
+    provideFirebaseApp(() => initializeApp(environment.firebase)),
+    provideAuth(() => getAuth()),
+    provideFirestore(() => getFirestore()),
+    provideStorage(() => getStorage()),
+    importProvidersFrom(HotToastModule.forRoot())
   ]
 }).catch(err => console.error(err));
